Wingflexer: Increase Update Frequency
Fixes the wingflexer causing wings to fall off / stretching wings unrealistically.
Previously, using <property-rule/>, the wingflexer updated with reference to the frame rate. This caused the wingflexer logic to break when the framerate is low, time warp is high or when a sudden lag spike occurs. By using <autopilot/>, this updates the wingflexer with reference to the FDM rate, or 120Hz, hence preventing issues that may occur with the wingflexer.